// src/integrations/supabase/types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      reviews: {
        Row: {
          created_at: string | null
          customer_name: string
          id: string
          is_featured: boolean | null
          rating: number | null
          review_text: string
        }
        Insert: {
          created_at?: string | null
          customer_name: string
          id?: string
          is_featured?: boolean | null
          rating?: number | null
          review_text: string
        }
        Update: {
          created_at?: string | null
          customer_name?: string
          id?: string
          is_featured?: boolean | null
          rating?: number | null
          review_text?: string
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      is_user_admin: {
        Args: { user_id: string }
        Returns: boolean
      }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}